intermediate changes ref:8e4ba2832950541c17c27ea93738c827c9c92309

+++ library/json/json_reader.cpp
@@ -406,7 +406,7 @@ namespace {
 
         template <class U>
         bool ProcessUint(U u) {
-            if (Y_LIKELY(u <= Max<int64_t>())) {
+            if (Y_LIKELY(u <= uint64_t(Max<int64_t>()))) {
                 return Impl.OnInteger(int64_t(u));
             } else {
                 return Impl.OnUInteger(u);
